Extracted `SparkKeyGenUtils`
